requirements.txt
numpy==1.20.1
tensorflow_gpu==1.15.0
Pillow==8.1.1
